update icon and something more

+++ core/database/src/main/java/com/billy/simpleunitconvert/core/database/UnitDatabase.kt
@@ -21,24 +21,37 @@ abstract class UnitDatabase : RoomDatabase() {
 
 val MIGRATION_1_2 = object : Migration(1, 2) {
     override fun migrate(database: SupportSQLiteDatabase) {
-        // Create a new table with the desired schema
-        database.execSQL("""
-            CREATE TABLE InformationEntity_new (
-                id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
-                uid TEXT
-            )
-        """)
+        // Check if the original table exists before attempting to migrate
+        val tableExists = database.query("SELECT name FROM sqlite_master WHERE type='table' AND name='InformationEntity'").count > 0
 
-        // Copy data from the old table to the new table
-        database.execSQL("""
-            INSERT INTO InformationEntity_new (id, uid)
-            SELECT id, uid FROM InformationEntity
-        """)
+        if (tableExists) {
+            // Create a new table with the desired schema
+            database.execSQL("""
+                CREATE TABLE IF NOT EXISTS InformationEntity_new (
+                    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+                    uid TEXT
+                )
+            """)
 
-        // Remove the old table
-        database.execSQL("DROP TABLE InformationEntity")
+            // Copy data from the old table to the new table
+            database.execSQL("""
+                INSERT INTO InformationEntity_new (id, uid)
+                SELECT id, uid FROM InformationEntity
+            """)
 
-        // Rename the new table to the original table name
-        database.execSQL("ALTER TABLE InformationEntity_new RENAME TO InformationEntity")
+            // Remove the old table
+            database.execSQL("DROP TABLE InformationEntity")
+
+            // Rename the new table to the original table name
+            database.execSQL("ALTER TABLE InformationEntity_new RENAME TO InformationEntity")
+        } else {
+            // If the original table doesn't exist, create the new table with the desired schema
+            database.execSQL("""
+                CREATE TABLE InformationEntity (
+                    id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL,
+                    uid TEXT
+                )
+            """)
+        }
     }
 }
